refactor enemy ai
