+++ 2024-1C-3051-fossers/src/GameObjects/Component.cs
@@ -3,8 +3,9 @@ using WarSteel.Scenes;
 
 namespace WarSteel.Entities;
 
-public interface IComponent {
-    void OnUpdate(GameObject self,GameTime gameTime, Scene scene);
+public interface IComponent
+{
+    void OnUpdate(GameObject self, GameTime gameTime, Scene scene);
     void OnStart(GameObject self, Scene scene);
     void Destroy(GameObject self, Scene scene);
 }
