+++ 2024-1C-3051-fossers/src/Scenes/SceneProcessors/WaveProcessor.cs
@@ -63,7 +63,7 @@ public class WaveProcessor : ISceneProcessor
         for (int i = 0; i < EnemiesLeft; i++)
         {
             Vector3 spawnPosition = VectorUtils.GetRandomVec3Pos(new(0, 100, 0), rand);
-            Enemy enemy = new Enemy(spawnPosition, WaveNumber * 5, _player);
+            Enemy enemy = new Enemy(spawnPosition, WaveNumber * 5, this);
             _scene.AddGameObject(enemy);
         }
     }
@@ -78,5 +78,6 @@ public class WaveProcessor : ISceneProcessor
     {
         EnemiesLeft--;
         EnemiesKilled++;
+        // play sound
     }
 }
